Allow `output` to accept procs as default
To fix a behaviour mismatch between `input` and `output` noted in
https://github.com/sunny/actor/issues/184, this PR allows `output`
to accept Procs (either with 0- or 1-arity) as defaults.

It's important to mention that output defaults will be evaluated (`call`'ed)
eagerly, i.e. at the same time the inputs (or non-proc output defaults)
are evaluated. It means that we will sometimes allocate new objects or
reference some constants for outputs which will be re-assigned during
actor's runtime. But I think that's consistent with how things worked before.

Another approach would be to assign non-procs output defaults before actor's
runtime (to preserve backward compact) and proc defaults after it is played.